backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2
llama-index==0.9.48
openai==1.12.0